revert to using xmalloc/xfree for allocations
I've (re)read our usage of them and observed no bad interactions
with the Ruby GC and no need to use RB_GC_GUARD anywhere.

xmalloc() does error-checking for us, so we'll raise
NoMemoryError instead of a segfaulting when/if malloc() fails to
release memory.

xmalloc() may increase the frequency of GC invocation in MRI
because it bumps the malloc counter (and xfree() cannot
decrement it), but MRI 1.9.3+ users can set a higher
RUBY_GC_MALLOC_LIMIT to reduce GC frequency (or improve GC :P).