reacquire GVL before retrying on EINTR on connect
If signals are received, we must reacquire the GVL to let
(Ruby-level) signal handlers run.  Otherwise, it may be
impossible to effectively cancel/interrupt a slow connection
negotiation.

Fortunately, connecting to MySQL servers is rare and rarely
interrupted (or even interruptible) by signal handlers, so
maybe nobody noticed this bug.