Moved/tidied tests. Bug fix in client.c
The bug fix in client.c is less than ideal but there isn't really
another option without some rearchitecting. Basically, wrapper->active
should not be set to 0 when mysql_use_result is being used until we
have iterated through the full result set (as the wrpaper is infact
still active). Unfortunately, we would need to change this within
Mysql2::Result, which isn't possible at the moment unless we pass the
client to the result object (which seems overkill). So, for now, the
query will fail with a "commands out of sync" error.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch master
# Your branch is ahead of 'origin/master' by 1 commit.
#
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	modified:   ext/mysql2/client.c
#	modified:   spec/mysql2/client_spec.rb
#	modified:   spec/mysql2/result_spec.rb
#
# Untracked files:
#   (use "git add <file>..." to include in what will be committed)
#
#	Gemfile.lock