Invert boolean casting logic
Change the logic of casting tinyints to boolean from == 1 being
true to != 0 being true. Whilst it's universally accepted that 0
is equivalent to false some database access libraries
(MS Access/ODBC for example) will write a -1 for a true value.
This is actually more consistent with Ruby behaviour where anything
other than nil or false is truthy.