Use string keys for test setup
When using symbol keys, this creates non deterministic behavior. It
creates a hash with both a symbol and string key for host, which means
that it depends on hash ordering which hosts it uses.

Since in 1.8 hash ordering is not defined, this creates random errors.

Fixes that sometimes it fails on Rubinius and REE in #334