Remove the check for SSL CA or Key being set.
Currently, the C bindings for the mysql2 gem enforce specifying either
the ssl-ca or ssl-key option, in order to set any of the SSL flags.
This makes the usage of mysql client worse for a few reasons. For one,
it's impossible to specify the ssl-capath flag by itself. This is
troubling if the server certificate is signed by a trusted authority,
whose certificate is present in /etc/ssl/certs, for example.

The other issue is that --ssl-cipher may not be specified on itsown.
The reason it may be desired is that it forces the client to use SSL
to connect to the server, without requiring the client to validate the
server certificate. In certain situations that may actually be
desirable (e.g. self-signed certificates which do not have a CA
certificate generated).

This patch also guards against calling mysql_ssl_set() uncondtionally
by ensuring that at least one of the SSL options is set.

Fixes: #355