Fix Timeout interrupt handling on Ruby 2.3 and protect Mysql2::Statement#execute
Timeout::ExitException was removed in Ruby 2.3.0, 2.2.3, and 2.1.8,
in favor of Timeout::Error. Backwards compatible aliases are provided
for Ruby 2.1.x and 2.2.x, but not earlier verions.

With thanks to @jeremy for PR #671 and @yui-knk for PR #677, this commit
also protects prepared statements from being interrupted, so the compat
shim is in Mysql2::Util.