change next_result to not block the interpreter
when actually preforming a multi-statement select with any non-trivial
components, the `next_result` call will block the ruby interpreter.
We fix this in two ways:

First, we wrap up the actual mysql_next_result call in a "no-gvl"
call so that the interpreter can do other work while next_result is
happening.  Second, we borrow a method from do_query and spin in a
select() loop on the mysql socket until the query is ready.  A quick
source dive into the underlying mysql source shows that next_result()
falls into the same path as mysql_real_query (cli_read_query_result), so
this should be a safe method.