Handle being terminated while reading query result (#811)
The problem is that if execution is terminated while reading the result
set, the next attempt to execute a query on the connection fails with
"This connection is still waiting for a result, try again once you have
the result". This error persists even if reconnect is enabled.

The issue stems from reading the result set using
rb_thread_call_without_gvl. If the calling thread is terminated inside
rb_thread_call_without_gvl, the function does not return and the caller
does not get a chance to reset the active thread associated with the
connection - which normally happens after the result set has been read.

The solution is to disconnect and mark the connection as inactive if the
thread is terminated while reading a result set.