Avoid invalidating re-assigned socket fd by fixing connected check (#847)
Previously the gem kept a field in the client's struct to keep track of
whether or not it is connected, but this could easily get out of sync with
the MYSQL struct. When invalidating the socket fd, it was getting the fd
from the MYSQL struct, which doesn't get it cleared in any way when
libmysqlclient closes the socket.  This means that invalidating the socket
fd can easily result in an unrelated fd being invalidated (e.g. another
connection)