Fix broken spec due to EXPLAIN EXTENDED on MySQL 8
The behavior that used to be triggered by EXPLAIN EXTENDED has been the
default for a long time and the previously deprecated EXTENDED keyword
was finally removed in MySQL 8.0.

To avoid breaking the spec on very old MySQL versions we just catch the
SQL syntax error and retry without the EXTENDED keyword.