Revert to precise on travis to fix mariadb install
The default dist is currently trusty, so the travis install scripts that
imply the old default of precise no longer work properly.

A better solution would be to update to the trusty (unsupported) or
xenial (supported) environment, but this should be the most minimal
change to restore the previous CI behavior.