Fix control flow in rb_set_ssl_mode_option for some client library versions
MySQL Connector/C 6.1.3 and above configure as HAVE_CONST_MYSQL_OPT_SSL_ENFORCE,
but their client library version is a different range than MySQL 5.7.3 - 5.7.10
that also have this option.

And generally there's a compiler warning here because of an incomplete control
flow if the library version doesn't match and exits the function without an
explicit return as required, so add a warning in this case.

Closes #1062