Add an option to use certifications generated from localhost.
In the Fedora project, we are running the mysql2 tests on the build environment
with a user permission, without root permission and without `sudo`.

In this case, we couldn't set up the custom domain "mysql2gem.example.com"
required to run SSL tests. This feature to create a set of the certifications
from the localhost gives an option to run the SSL tests executed in the environment.

How to generate the certificaton files:

```
$ cd spec/ssl/
$ bash gen_certs.sh
```

The files are generated in the `spec/ssl/<host name>`. Added the new files
generated from the domain localhost in the `spec/ssl/localhost`.

How to use:

```
$ TEST_RUBY_MYSQL2_SSL_CERT_HOST=localhost \
  bundle exec rake spec
```