CI: Set the verbose option in the Makefile to print compiling commands.
Enable the verbose option in the Makefile used to compile the extension to
print the compiling command lines in the log.

Note that we use MAKEFLAGS rather than GNUMAKEFLAGS[1], because the
GNUMAKEFLAGS doesn't work for the make used in CI MacOS cases.

We don't need to set the verbose option to the CI Fedora and CentOS cases
(`.github/workflows/container.yml`). Because the Rubies from the Ruby RPM
package used in Fedora and CentOS cases are already enabling the verbose option.

[1] https://www.gnu.org/software/make/manual/html_node/Options_002fRecursion.html