Improve prepared statement ergonomics (#1385)
While working on Mysql2 prepared statement support in Rails,
I found them very hard to use. Most notably we sometimes need to
close them eagerly, but it's complicated by the fact that you can
only call `close` once, any subsequent call raises an error. There
was also no way to check if a statement was closed or not.

This change noops on repeat calls to `close` and adds `closed?`