CI: Add an option to use SSL certifications generated from specific host. (#1310)
In the Fedora project, we are running the mysql2 tests on the build environment
with a user permission, without root permission and without `sudo`.

In this case, we couldn't set up the custom domain "mysql2gem.example.com"
to run SSL tests. The feature to create a set of the SSL certifications
from the localhost gives an option to run the SSL tests executed in the environment.

How to generate the certificaton files:

```
$ cd spec/ssl/
$ TEST_RUBY_MYSQL2_SSL_CERT_HOST=localhost bash gen_certs.sh
```

The files are generated in the `spec/ssl` directory.

How to use:

```
$ TEST_RUBY_MYSQL2_SSL_CERT_HOST=localhost \
bundle exec rake spec
```